Use appropriate digest algorithm during signature creation
Pass the public key instead of the marshalled public key to
`digestAlgorithmForPublicKey` in `SignedData.AddSignerInfo`.

Previously, the marshalled public key was passed instead of the actual
public key. The result is that always SHA256 was being selected, even
for ECDSA where the hash algorithm should be selected based on the curve.